pyproject.toml
[tool.poetry]
name = "learning-platform-drf"
version = "0.1.0"
description = ""
authors = ["vlad <sapozhnikov.061098@mail.ru>"]
readme = "README.md"
packages = [{include = "learning_platform_drf"}]

[tool.poetry.dependencies]
python = "^3.11"
django = "^4.2.4"
psycopg2-binary = "^2.9.7"
pillow = "^10.0.0"
djangorestframework = "^3.14.0"
python-dotenv = "^1.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
